Fixes for Deployment with Mock-DFSP
Contains 2 docker-compose files. Rename as appropriate for the deployment required. As is, the Simulator with the Scheme Adapter ONLY, will deploy. Swop the docker-compose-sim-dfsp in, as the current docker-compose.yml file and the Simulator with the Scheme Adapter as Payee, together with the Mock-DFSP and its Scheme Adapter, will deploy