Mostly fixing inter-service communication
Also added some config options that exist in the sim but weren't exposed
in the chart. Commented out some config and added commentary to indicate
why it shouldn't be specified. These items have been specified in
templates. The user should still be able to override them should they
really wish to do so.